Resolve "Meta Collector -> New Modi"

Closes #15

See merge request ddsg/ml4sys/local-cardinality-estimation!11

+++ meta_collector/meta_collector.py
@@ -1,8 +1,22 @@
+import os.path
+from enum import Enum
 from typing import List, Tuple, Dict
 
 import psycopg2 as postgres
 import yaml
 from sklearn.preprocessing import LabelEncoder
+
+
+class CreationMode(Enum):
+    """
+    Enum for the different possibilities to use the MetaCollector.
+
+    0 -> don't create table, 1 -> create temporary table, 2 -> create permanent table
+    """
+
+    NONE = 0
+    TEMPORARY = 1
+    PERMANENT = 2
 
 
 class MetaCollector:
@@ -88,7 +102,9 @@ class MetaCollector:
             self.conn.close()
 
     def setup_view(self, table_names: List[str or Tuple[str, str]], columns: List[str],
-                   join_atts: List[str or Tuple[str, str]] = None, cube: bool = False) -> (List[Tuple[str, str]], int):
+                   join_atts: List[str or Tuple[str, str]] = None, cube: bool = False,
+                   mode: CreationMode = CreationMode.NONE) -> (
+            List[Tuple[str, str]], int):
         """
         Create the tables tmpview and if cube==True also tmpview_cube containing the metadata for the given tables
         joined on the attributes and projected on the columns.
@@ -100,6 +116,7 @@ class MetaCollector:
             join the tables on. -> is optional, because there is no join if there is only one table and so there would
             be no join-attribute needed in that case
         :param cube: boolean whether to create the *_cube table, too
+        :param mode: see CreationMode-Enum
         :return: first: a list of tuples containing the name and the datatype for the columns, each as string
             second: the maximal cardinality as integer
         """
@@ -108,11 +125,10 @@ class MetaCollector:
             raise ConnectionError("The database-connection may not have been initialized correctly. Make sure to call "
                                   "'open_database_connection' before this method.")
 
-        # drops maybe already existing tables with metadata
-        sql = """DROP TABLE IF EXISTS tmpview; DROP TABLE IF EXISTS tmpview_cube;"""
-        if self.debug:
-            print("Executing: {}".format(sql))
-        self.cur.execute(sql)
+        if mode == CreationMode.TEMPORARY:
+            # drops maybe already existing tables with metadata
+            sql = """DROP TABLE IF EXISTS tmpview; DROP TABLE IF EXISTS tmpview_cube;"""
+            self.execute_sql(sql)
 
         # get column-name and datatype for the requested columns of the corresponding tables
         if isinstance(table_names[0], list) or isinstance(table_names[0], tuple):
@@ -131,17 +147,17 @@ class MetaCollector:
                 sql = """SELECT column_name, data_type FROM information_schema.columns 
                                      WHERE table_schema = 'public' AND table_name IN ('{}') 
                                      AND column_name IN ('{}') ORDER BY 1;""".format(table_name, columns_string)
-            if self.debug:
-                print("Executing: {}".format(sql))
-            self.cur.execute(sql)
+            self.execute_sql(sql)
 
             # get all remaining rows from the result set
             column_type = self.cur.fetchall()
             for column in column_type:
                 if isinstance(table_name, list) or isinstance(table_name, tuple):
-                    columns_types.append((column[0], table_name[1], column[1]))
+                    min_max_step, encoders = self.collect_meta(table_name[0], (column[0], column[1]))
+                    columns_types.append((column[0], table_name[1], column[1], min_max_step, encoders))
                 else:
-                    columns_types.append((column[0], table_name, column[1]))
+                    min_max_step, encoders = self.collect_meta(table_name, (column[0], column[1]))
+                    columns_types.append((column[0], table_name, column[1], min_max_step, encoders))
 
         # create table with the tuples for the given tables (join-result if
         # more than one table) with projection on the given columns
@@ -159,19 +175,50 @@ class MetaCollector:
                 attributes_string = " AND ".join(["t{}.{} = t{}.{}".format(1, join[0], i + 2, join[1]) for i, join in
                                                   enumerate(join_atts)])
 
-            sql = """CREATE TABLE tmpview AS (SELECT {} FROM {} WHERE {});""".format(columns_string, tables_string,
-                                                                                     attributes_string)
+            if mode == CreationMode.NONE:
+                sql = """ SELECT COUNT(*) FROM {} WHERE {} """.format(tables_string, attributes_string)
+            elif mode == CreationMode.TEMPORARY:
+                new_table_name = "tmpview"
+                sql = """CREATE TABLE {} AS (SELECT {} FROM {} WHERE {});""".format(new_table_name, columns_string,
+                                                                                    tables_string, attributes_string)
+            elif mode == CreationMode.PERMANENT:
+                new_table_name = "tmpview_{}".format("_".join(table[0] for table in table_names))
+                sql = """CREATE TABLE IF NOT EXISTS {} AS (SELECT {} FROM {} WHERE {});""".format(new_table_name,
+                                                                                                  columns_string,
+                                                                                                  tables_string,
+                                                                                                  attributes_string)
+            else:
+                raise ValueError("Invalid mode selected. There are only 0, 1 or 2 as modes available!")
         else:
             columns_string = ",".join(["coalesce({col},'-1') AS {col}".format(col=col[0]) if "character" in col[2]
                                        else "{col}".format(col=col[0]) for col in columns_types])
-            sql = """CREATE TABLE tmpview AS (SELECT {} FROM {});""".format(columns_string, table_names[0])
 
-        if self.debug:
-            print("Executing: {}".format(sql))
-        self.cur.execute(sql)
+            if mode == CreationMode.NONE:
+                sql = """ SELECT COUNT(*) FROM {} """.format(table_names[0])
+            elif mode == CreationMode.TEMPORARY:
+                new_table_name = "tmpview"
+                sql = """CREATE TABLE {} AS (SELECT {} FROM {});""".format(new_table_name, columns_string,
+                                                                           table_names[0])
+            elif mode == CreationMode.PERMANENT:
+                new_table_name = "tmpview_{}".format(table_names[0])
+                sql = """CREATE TABLE IF NOT EXISTS {} AS (SELECT {} FROM {});""".format(new_table_name, columns_string,
+                                                                                         table_names[0])
+            else:
+                raise ValueError("Invalid mode selected. There are only 0, 1 or 2 as modes available!")
+
+        if mode == CreationMode.TEMPORARY or mode == CreationMode.PERMANENT:
+            # self.execute_sql(sql)
+            # sql = """ANALYZE {};""".format(new_table_name)
+            self.execute_sql(sql)
+            sql = """SELECT count(*) FROM {};""".format(new_table_name)
+
+        self.execute_sql(sql)
+
+        # get the count of tuples in the tmpview
+        max_card = self.cur.fetchall()[0]
 
         if cube:
-            sql = """CREATE TABLE tmpview_cube AS (SELECT {col}, count(*)::integer, 0.0 as perc FROM tmpview 
+            sql = """CREATE TABLE tmpview_cube AS (SELECT {col}, count(*)::integer, 0.0 as perc FROM tmpview
                 GROUP BY GROUPING SETS(({col})));
                 UPDATE tmpview_cube SET perc = count/(SELECT SUM(count) FROM tmpview_cube);""".format(
                 col=",".join(column.split(".")[-1] for column in columns))
@@ -179,28 +226,20 @@ class MetaCollector:
                 print("Executing: {}".format(sql))
             self.cur.execute(sql)
 
-            sql = """ANALYZE tmpview; ANALYZE tmpview_cube;"""
-        else:
-            sql = """ANALYZE tmpview;"""
+            sql = """ANALYZE tmpview_cube;"""
 
-        if self.debug:
-            print("Executing: {}".format(sql))
-        self.cur.execute(sql)
-
-        # get the count of tuples in the tmpview
-        sql = """SELECT count(*) FROM tmpview;"""
-        if self.debug:
-            print("Executing: {}".format(sql))
-        self.cur.execute(sql)
-        max_card = self.cur.fetchall()[0]
+            if self.debug:
+                print("Executing: {}".format(sql))
+            self.cur.execute(sql)
 
         return columns_types, max_card
 
-    def collect_meta(self, columns: List[Tuple[str, str, str]]) -> (Dict[str, Tuple[int, int, int]], Dict):
+    def collect_meta(self, tablename: str, column: Tuple[str, str]) -> (Tuple[int, int, int], Dict):
         """
         after execution of setup_view this function returns the min and max values for the meta-table and the encoders
 
-        :param columns: a list of tuples containing the name and the datatype for the columns, each as string
+        :param tablename: String containing the name of the table where to find the column
+        :param column: a tuple containing the name and the datatype for the column, each as string
         :return: first: dictionary with the attribute-name as key and a tuple containing min-value, max-value and
             step-size (all as int) as value
             second: a dictionary of the not integer encoders with key attribute-name and value the encoder
@@ -208,30 +247,31 @@ class MetaCollector:
 
         if not self.cur:
             raise ConnectionError("The database-connection may not have been initialized correctly. Make sure to call "
-                                  "'open_database_connection' befrore this method.")
+                                  "'open_database_connection' before this method.")
 
-        min_max = {}
         encoders = {}
 
-        for col in columns:
-            sql = """SELECT {col}, count(*) from tmpview GROUP BY {col};""".format(col=col[0])
+        if column[1] != "integer":
+            sql = """SELECT {col}, count(*) from {tab} GROUP BY {col};""".format(col=column[0], tab=tablename)
 
-            if self.debug:
-                print("Executing: {}".format(sql))
-            self.cur.execute(sql)
+            self.execute_sql(sql)
             tmp = self.cur.fetchall()
 
             cats = [x[0] for x in tmp if x[0] is not None]
 
-            if col[2] != "integer":
-                le = LabelEncoder()
-                cats = le.fit_transform(sorted(cats))
+            le = LabelEncoder()
+            cats = le.fit_transform(sorted(cats))
 
-                encoders[col[0]] = le
+            encoders[column[0]] = le
 
-            min_max[col[0]] = (min(cats), max(cats), 1)
+            return (min(cats), max(cats), 1), encoders
+        else:
+            sql = """SELECT MIN({col}) AS min, MAX({col}) AS max FROM {tab}""".format(col=column[0], tab=tablename)
 
-        return min_max, encoders
+            self.execute_sql(sql)
+            tmp = self.cur.fetchall()
+
+            return (tmp[0][0], tmp[0][1], 1), encoders
 
     def eliminate_duplicates(self, columns: List[str]) -> List[str]:
         """
@@ -254,7 +294,7 @@ class MetaCollector:
 
     def get_meta(self, table_names: List[str or Tuple[str, str]], columns: List[str],
                  join_atts: List[str or Tuple[str, str]] = None, save: bool = True, save_file_name: str = None,
-                 batchmode: bool = False) -> Dict:
+                 batchmode: bool = False, mode: CreationMode = CreationMode.NONE) -> Dict:
         """
         Method for the whole process of collecting the meta-information for the given tables joined on the given
         attributes and projected on the given columns.
@@ -269,21 +309,21 @@ class MetaCollector:
         :param save_file_name: name for the save-file for the meta_information -> not needed if save==False
         :param batchmode: whether the meta data is collected in batches or not -> connection to db held open if batch
             mode
+        :param mode: see CreationMode-Enum
         :return: dictionary containing the meta-information
         """
 
         if not batchmode:
             self.open_database_connection()
 
-        cols, max_card = self.setup_view(table_names, columns, join_atts, True)
-        mm, encs = self.collect_meta(cols)
+        cols, max_card = self.setup_view(table_names, columns, join_atts, cube=False, mode=mode)
 
         already_seen = []
         for index, col in enumerate(cols):
             if col[0] not in already_seen:
                 already_seen.append(col[0])
             else:
-                cols[index] = (col[0], col[1], col[2], "_".join([col[1], col[0]]))
+                cols[index] = (col[0], col[1], col[2], col[3], col[4], "_".join([col[1], col[0]]))
 
         if not batchmode:
             self.close_database_connection()
@@ -291,8 +331,6 @@ class MetaCollector:
         result_dict = {"table_names": table_names,
                        "columns": cols,
                        "join_attributes": join_atts,
-                       "min_max_step": mm,
-                       "encodings": encs,
                        "max_card": max_card}
 
         if not batchmode:
@@ -306,7 +344,8 @@ class MetaCollector:
 
         return result_dict
 
-    def get_meta_from_file(self, file_path: str, save: bool = True, save_file_name: str = None) -> Dict[int, any]:
+    def get_meta_from_file(self, file_path: str, save: bool = True, save_file_name: str = None,
+                           mode: CreationMode = CreationMode.NONE, override: bool = True) -> Dict[int, any]:
         """
         Method for collecting meta data for the information given in a file from Crawler or at least a file formatted
         like this.
@@ -314,8 +353,18 @@ class MetaCollector:
         :param file_path: Path to the file. Format has to be the same like the output of Crawler
         :param save: Whether to save the information to file or not. -> It is recommended to do so.
         :param save_file_name: Optional name for the file.
+        :param mode: see CreationMode-Enum
+        :param override: Whether to override an already existing meta_information file.
         :return: The solution dict.
         """
+
+        if override:
+            if save_file_name:
+                if os.path.isfile(save_file_name + ".yaml"):
+                    os.remove(save_file_name + ".yaml")
+            else:
+                if os.path.isfile("meta_information.yaml"):
+                    os.remove("meta_information.yaml")
 
         solution_dict = {}
 
@@ -328,7 +377,7 @@ class MetaCollector:
             solution_dict = {index: self.get_meta(table_names=batch[index]["table_names"],
                                                   columns=batch[index]["selection_attributes"],
                                                   join_atts=batch[index]["join_attributes"],
-                                                  save=False, batchmode=True)}
+                                                  save=False, batchmode=True, mode=mode)}
 
             if save:
                 if save_file_name:
@@ -348,15 +397,33 @@ class MetaCollector:
         :param file_name: the name (without file-type) for the save-file
         :param mode: The mode to open the file. Some common possibilities are 'w', 'w+', 'r', 'a', 'a+'
         """
-        
+
         if self.debug:
             print("Saving: {} to {}".format(meta_dict, (file_name + ".yaml") if file_name else "file"))
 
         with open(file_name + ".yaml", mode) as file:
             yaml.safe_dump(meta_dict, file)
 
+    def execute_sql(self, sql_string: str):
+        """
+        Method for executing a SQL-Query.
+
+        :param sql_string: The SQL-Query to execute
+        """
+
+        if not sql_string:
+            raise ValueError("The SQL-String must not be empty!")
+        if not self.cur:
+            raise ConnectionError("The database-connection may not have been initialized correctly. Make sure to call "
+                                  "'open_database_connection' before this method.")
+
+        if self.debug:
+            print("Executing: {}".format(sql_string))
+        self.cur.execute(sql_string)
+
 
 mc = MetaCollector()
 # example which should work -> takes quite a while to be processed
 # mc.get_meta(["title", "cast_info"], ["kind_id", "person_id", "role_id"], [("id", "movie_id")])
-mc.get_meta_from_file("../crawler/solution_dict.yaml")
+
+mc.get_meta_from_file(file_path="../assets/solution_dict.yaml", save_file_name="../assets/meta_information")
